feat(pieces): refactor ConnectUC triggers with reusable dropdown props
Create reusable property functions for domain and users dropdowns:
- Add src/lib/common/props.ts with domainProp() and usersProp()
- Refactor new-recording trigger to use reusable props
- Add dynamic dropdowns to new-call-transcription trigger
- Domain: single-select dropdown from /activepieces/domains
- Users: multi-select dropdown from /activepieces/subscribers
- Users dropdown includes "All Always" option with value "*"

Benefits: DRY principle, consistent UX, easier maintenance
