+++ packages/pieces/community/connectuc/src/lib/triggers/new-recording.ts
@@ -1,8 +1,7 @@
-import { createTrigger, TriggerStrategy, OAuth2PropertyValue, Property } from '@activepieces/pieces-framework';
+import { createTrigger, TriggerStrategy, OAuth2PropertyValue } from '@activepieces/pieces-framework';
 import { connectucAuth } from '../../index';
 import { registerConnectUCWebhook, unregisterConnectUCWebhook } from '../common/webhook-helpers';
-import { connectucApiCall } from '../common/api-helpers';
-import { HttpMethod } from '@activepieces/pieces-common';
+import { domainProp, usersProp } from '../common/props';
 
 export const newRecording = createTrigger({
     auth: connectucAuth,
@@ -10,113 +9,8 @@ export const newRecording = createTrigger({
     displayName: 'New Recording',
     description: 'Triggers when a new call recording is available',
     props: {
-        domain: Property.Dropdown({
-            displayName: 'Domain',
-            description: 'Select domain to which this trigger applies',
-            required: false,
-            auth: connectucAuth,
-            refreshers: [],
-            options: async ({ auth }) => {
-                if (!auth) {
-                    return {
-                        disabled: true,
-                        placeholder: 'Please connect your account first',
-                        options: [],
-                    };
-                }
-
-                try {
-                    const authValue = auth as OAuth2PropertyValue;
-
-                    interface DomainInfo {
-                        domain: string;
-                        reseller: string;
-                        description: string;
-                    }
-
-                    const domainsResponse = await connectucApiCall<Record<string, DomainInfo>>({
-                        accessToken: authValue.access_token,
-                        endpoint: '/activepieces/domains',
-                        method: HttpMethod.GET,
-                    });
-
-                    const options = Object.values(domainsResponse).map(domainInfo => ({
-                        label: `${domainInfo.description} (${domainInfo.domain})`,
-                        value: domainInfo.domain,
-                    }));
-
-                    return {
-                        disabled: false,
-                        options,
-                    };
-                } catch (error) {
-                    console.error('Error fetching domains:', error);
-                    return {
-                        disabled: true,
-                        placeholder: 'Error loading domains',
-                        options: [],
-                    };
-                }
-            },
-        }),
-        users: Property.MultiSelectDropdown({
-            displayName: 'Users',
-            description: 'Select users to which this trigger applies',
-            required: false,
-            auth: connectucAuth,
-            refreshers: [],
-            options: async ({ auth }) => {
-                if (!auth) {
-                    return {
-                        disabled: true,
-                        placeholder: 'Please connect your account first',
-                        options: [],
-                    };
-                }
-
-                try {
-                    const authValue = auth as OAuth2PropertyValue;
-
-                    interface Subscriber {
-                        first_name: string;
-                        last_name: string;
-                        user: string;
-                    }
-
-                    const subscribers = await connectucApiCall<Subscriber[]>({
-                        accessToken: authValue.access_token,
-                        endpoint: '/activepieces/subscribers',
-                        method: HttpMethod.GET,
-                    });
-
-                    const options = [
-                        {
-                            label: 'All Always',
-                            value: '*',
-                        },
-                        ...subscribers.map(subscriber => {
-                            const fullName = `${subscriber.first_name} ${subscriber.last_name}`.trim();
-                            return {
-                                label: `${fullName} (${subscriber.user})`,
-                                value: subscriber.user,
-                            };
-                        })
-                    ];
-
-                    return {
-                        disabled: false,
-                        options,
-                    };
-                } catch (error) {
-                    console.error('Error fetching subscribers:', error);
-                    return {
-                        disabled: true,
-                        placeholder: 'Error loading subscribers',
-                        options: [],
-                    };
-                }
-            },
-        }),
+        domain: domainProp(),
+        users: usersProp(),
     },
     sampleData: {
         dateTime: '2025-11-06T14:07:59.000Z',
